- set project finish sudah bisa - set user category prefrence dah done - post rating juga dah aman mas

+++ Aplikasi/app/src/main/java/com/capstone/idekita/response/GetAllProjectResponse.kt
@@ -59,11 +59,11 @@ data class ProjectsItem(
     val jumlahRaters: Int,
 
     @field:SerializedName("mean_rate")
-    val meanRate: Int,
+    val meanRate: Double,
 
 
 
-):Parcelable
+    ):Parcelable
 
 @Parcelize
 data class Category(
